+++ tests/Feature/EventTrackingModelEventsTest.php
@@ -4,6 +4,7 @@ use Faker\Factory;
 use Illuminate\Foundation\Auth\User;
 use MOIREI\EventTracking\Channels\GoogleAnalytics;
 use MOIREI\EventTracking\Channels\Mixpanel;
+use MOIREI\EventTracking\Channels\Posthog;
 use MOIREI\EventTracking\EventTrackingServiceProvider;
 use MOIREI\EventTracking\Facades\Events;
 
@@ -16,18 +17,20 @@ beforeEach(function () {
 });
 
 it('should track model events', function () {
-    /** @var \Mockery\LegacyMockInterface */
     $eventChannel1 = \Mockery::mock(GoogleAnalytics::class.'[track,initialize]');
-    /** @var \Mockery\LegacyMockInterface */
     $eventChannel2 = \Mockery::mock(Mixpanel::class.'[track,initialize]');
+    $eventChannel3 = \Mockery::mock(Posthog::class.'[track,initialize]');
 
     $eventChannel1->shouldReceive('track');
     $eventChannel1->shouldReceive('initialize');
     $eventChannel2->shouldReceive('track');
     $eventChannel2->shouldReceive('initialize');
+    $eventChannel3->shouldReceive('track');
+    $eventChannel3->shouldReceive('initialize');
 
     $this->instance(GoogleAnalytics::class, $eventChannel1);
     $this->instance(Mixpanel::class, $eventChannel2);
+    $this->instance(Posthog::class, $eventChannel3);
     app()->register(EventTrackingServiceProvider::class);
 
     Events::observe(User::class);