+++ tests/Feature/EventTrackingTest.php
@@ -9,27 +9,27 @@ use MOIREI\EventTracking\Facades\Events;
 uses()->group('event-tracking');
 
 it('should send event to all channels', function () {
-    /** @var \Mockery\LegacyMockInterface */
     $eventChannel1 = \Mockery::mock(GoogleAnalytics::class.'[track,initialize]');
-    /** @var \Mockery\LegacyMockInterface */
     $eventChannel2 = \Mockery::mock(Mixpanel::class.'[track,initialize]');
+    $eventChannel3 = \Mockery::mock(PostHog::class.'[track,initialize]');
 
     $eventChannel1->shouldReceive('track');
     $eventChannel1->shouldReceive('initialize');
     $eventChannel2->shouldReceive('track');
     $eventChannel2->shouldReceive('initialize');
+    $eventChannel3->shouldReceive('track');
+    $eventChannel3->shouldReceive('initialize');
 
     $this->instance(GoogleAnalytics::class, $eventChannel1);
     $this->instance(Mixpanel::class, $eventChannel2);
+    $this->instance(PostHog::class, $eventChannel3);
     app()->register(EventTrackingServiceProvider::class);
 
     Events::track('my-event');
 });
 
 it('should send event to GA channel', function () {
-    /** @var \Mockery\LegacyMockInterface */
     $eventChannel1 = \Mockery::mock(GoogleAnalytics::class.'[track,initialize]');
-    /** @var \Mockery\LegacyMockInterface */
     $eventChannel2 = \Mockery::mock(Mixpanel::class.'[track,initialize]');
     $eventChannel3 = \Mockery::mock(PostHog::class.'[track,initialize]');